+++ modules/dnn/src/torch/torch_importer.cpp
@@ -258,7 +258,7 @@ struct TorchImporter : public ::cv::dnn::Importer
         case CV_USRTYPE1:
         {
             double *buf = storageMat.ptr<double>();
-            THFile_readLongRaw(file, (long*)buf, size);
+            THFile_readLongRaw(file, (long long*)buf, size);
 
             for (size_t i = (size_t)size; i-- > 0; )
                 buf[i] = ((long*)buf)[i];
@@ -365,8 +365,8 @@ struct TorchImporter : public ::cv::dnn::Importer
     void readTorchTensor(int indexTensor, int typeTensor)
     {
         int ndims = readInt();
-        AutoBuffer<long, 4> sizes(ndims);
-        AutoBuffer<long, 4> steps(ndims);
+        AutoBuffer<long long, 4> sizes(ndims);
+        AutoBuffer<long long, 4> steps(ndims);
         THFile_readLongRaw(file, sizes, ndims);
         THFile_readLongRaw(file, steps, ndims);
         long offset = readLong() - 1;
